Clean up usage and quota
Give user agents even more freedom in determining the usage and quota
fields for an origin and stop trying to define the global environment
in which they fit.

Also, make it more clear that each origin has a single site storage
unit which has a single box.

Also, indicate that credentials storage should not be grouped with
site/network storage in UI as if it were all a single thing.

Fixes #22.